requirements.txt
openai-whisper
opencv-python
python-pptx
ffmpeg-python
transformers
pillow